Move load data to global variables
Reduce runtime by storing data read from files.
Don't need to use File IO for each function call

Is a tad messy in terms of number of variables. Can refactor into a dictionary if wanted. Also not sure what naming convention is for 'private' variables that are not in a class.